Create FIT.py
This solution is a part of:
    June Long Two 2022 Division 4 (Rated)

Problem Statement:
Chef wants to become fit for which he decided to walk to the office and return home by walking. It is known that Chef's office is X km away from his home.

If his office is open on 5 days in a week, find the number of kilometers Chef travels through office trips in a week.